Modifications And Bug Fixes

+++ Backend/src/main/java/com/example/backend/project/controller/ProjectController.java
@@ -1,7 +1,9 @@
 package com.example.backend.project.controller;
 
+import com.example.backend.collaborations.model.Collaboration;
 import com.example.backend.collaborations.repo.CollaborationsRepository;
 import com.example.backend.collaborations.service.CollaborationsService;
+import com.example.backend.companies.service.CompanyService;
 import com.example.backend.project.controller.dpo.FRTeamMemberDPO;
 import com.example.backend.project.controller.dpo.ProjectAndFRTeamMembersDPO;
 import com.example.backend.project.controller.dto.CollaborationDTO;
@@ -28,6 +30,7 @@ public class ProjectController {
     private final ProjectService projectService;
     private final UserService userService;
     private final CollaborationsService collaborationsService;
+    private final CompanyService companyService;
 
     @GetMapping()
     @ResponseBody
@@ -49,6 +52,7 @@ public class ProjectController {
         if (userService.findByEmail(email) == null)
             return new ResponseEntity("You don't have access to CDB", HttpStatus.UNAUTHORIZED);
 
+        if (!projectService.existsById(id)) return new ResponseEntity("Project not found", HttpStatus.NOT_FOUND);
         Project project = projectService.findById(id);
         Set<FRTeamMemberDPO> frTeamMemberDPOSet = new HashSet<>();
         for (AppUser appUser: project.getFrteammembers()){ //iterating through project's FR team members to extract id, first and last name
@@ -103,6 +107,7 @@ public class ProjectController {
         if (!a.contains(userService.findByEmail(email).getAuthority()))
             return new ResponseEntity("You don't have premission to this resource", HttpStatus.UNAUTHORIZED);
 
+        if (!projectService.existsById(id)) return new ResponseEntity("Project not found", HttpStatus.NOT_FOUND);
         return new ResponseEntity(projectService.updateProject(projectDTO, id), HttpStatus.OK);
     }
 
@@ -117,22 +122,25 @@ public class ProjectController {
         if (!a.contains(userService.findByEmail(email).getAuthority()))
             return new ResponseEntity("You don't have premission to this resource", HttpStatus.UNAUTHORIZED);
 
+        if (!projectService.existsById(id)) return new ResponseEntity("Project not found", HttpStatus.NOT_FOUND);
+        projectService.deleteProject(id);
         return new ResponseEntity("Project under id: " + id + " Successfully deleted", HttpStatus.OK);
     }
 
     @GetMapping("/{id}/collaborations")
     public ResponseEntity getCollaborations(@RequestHeader String googleTokenEncoded, @PathVariable Long id){
-        List<AUTHORITY> a = List.of(AUTHORITY.MODERATOR, AUTHORITY.ADMIN);
         String email = JwtVerifier.verifyAndReturnEmail(googleTokenEncoded);
         if (email == null)
             return new ResponseEntity("Token is missing or invalid", HttpStatus.UNAUTHORIZED);
         if (userService.findByEmail(email) == null)
             return new ResponseEntity("You don't have access to CDB", HttpStatus.UNAUTHORIZED);
-        AppUser user = userService.findByEmail(email);
+        if (!projectService.existsById(id))
+            return new ResponseEntity("Project not found", HttpStatus.NOT_FOUND);
+        AppUser endpointCaller = userService.findByEmail(email);
         Project project = projectService.findById(id);
-        if (user.getAuthority() == AUTHORITY.OBSERVER ||
-                (user.getAuthority() == AUTHORITY.FR_TEAM_MEMBER && !project.getFrteammembers().contains(user)) ||
-                (user.getAuthority() == AUTHORITY.FR_RESPONSIBLE && project.getFRResp() != user)) {
+        if (endpointCaller.getAuthority() == AUTHORITY.OBSERVER ||
+                (endpointCaller.getAuthority() == AUTHORITY.FR_TEAM_MEMBER && !project.getFrteammembers().contains(endpointCaller)) ||
+                (endpointCaller.getAuthority() == AUTHORITY.FR_RESPONSIBLE && project.getFRResp().getId() != endpointCaller.getId())) {
             return new ResponseEntity("You don't have premission to this resource", HttpStatus.UNAUTHORIZED);
         }
 
@@ -147,10 +155,11 @@ public class ProjectController {
             return new ResponseEntity("Token is missing or invalid", HttpStatus.UNAUTHORIZED);
         if (userService.findByEmail(email) == null)
             return new ResponseEntity("You don't have access to CDB", HttpStatus.UNAUTHORIZED);
+        if (!projectService.existsById(projectid)) return new ResponseEntity("Project not found", HttpStatus.NOT_FOUND);
         AppUser user = userService.findByEmail(email);
         Project project = projectService.findById(projectid);
         if (user.getAuthority() == AUTHORITY.OBSERVER ||
-                (user.getAuthority() == AUTHORITY.FR_RESPONSIBLE && project.getFRResp() != user)) {
+                (user.getAuthority() == AUTHORITY.FR_RESPONSIBLE && project.getFRResp().getId() != user.getId())) {
             return new ResponseEntity("You don't have premission to this resource", HttpStatus.UNAUTHORIZED);
         }
 
@@ -160,27 +169,33 @@ public class ProjectController {
     @PutMapping("/{projectid}/collaborations/{companyid}")
     public ResponseEntity addCollaboration(@RequestHeader String googleTokenEncoded, @PathVariable Long projectid,
                                            @PathVariable Long companyid, @RequestBody CollaborationDTO collaborationDTO) {
-        List<AUTHORITY> a = List.of(AUTHORITY.MODERATOR, AUTHORITY.ADMIN);
         String email = JwtVerifier.verifyAndReturnEmail(googleTokenEncoded);
         if (email == null)
             return new ResponseEntity("Token is missing or invalid", HttpStatus.UNAUTHORIZED);
         if (userService.findByEmail(email) == null)
             return new ResponseEntity("You don't have access to CDB", HttpStatus.UNAUTHORIZED);
+        if (!projectService.existsById(projectid))
+            return new ResponseEntity("Project not found", HttpStatus.NOT_FOUND);
+        if (!companyService.existsById(companyid))
+            return new ResponseEntity("Company not found", HttpStatus.NOT_FOUND);
+
         AppUser user = userService.findByEmail(email);
         Project project = projectService.findById(projectid);
         if (user.getAuthority() == AUTHORITY.OBSERVER ||
-                (user.getAuthority() == AUTHORITY.FR_TEAM_MEMBER && collaborationsService.getContactRespByCollaboratonId(projectid, companyid) != user ||
-                (user.getAuthority() == AUTHORITY.FR_RESPONSIBLE && project.getFRResp() != user))) {
+                (user.getAuthority() == AUTHORITY.FR_TEAM_MEMBER && collaborationsService.getCollaborationByCollaborationId(projectid, companyid).getContactResponsible().getId() != user.getId() ||
+                (user.getAuthority() == AUTHORITY.FR_RESPONSIBLE && project.getFRResp().getId() != user.getId()))) {
             return new ResponseEntity("You don't have premission to this resource", HttpStatus.UNAUTHORIZED);
         }
 
-        return new ResponseEntity(collaborationsService.updateCollaboration(projectid, companyid, collaborationDTO), HttpStatus.OK);
+        Collaboration collaboration = collaborationsService.updateCollaboration(projectid, companyid, collaborationDTO);
+        if (collaboration == null) return new ResponseEntity("Collaboration not found", HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity(collaboration, HttpStatus.OK);
     }
 
     @DeleteMapping("/{projectid}/collaborations/{companyid}")
     public ResponseEntity deleteCollaboration(@RequestHeader String googleTokenEncoded, @PathVariable Long projectid,
-                                           @PathVariable Long companyid, @RequestBody CollaborationDTO collaborationDTO) {
-        List<AUTHORITY> a = List.of(AUTHORITY.MODERATOR, AUTHORITY.ADMIN);
+                                           @PathVariable Long companyid) {
         String email = JwtVerifier.verifyAndReturnEmail(googleTokenEncoded);
         if (email == null)
             return new ResponseEntity("Token is missing or invalid", HttpStatus.UNAUTHORIZED);
@@ -189,7 +204,7 @@ public class ProjectController {
         AppUser user = userService.findByEmail(email);
         Project project = projectService.findById(projectid);
         if (user.getAuthority() == AUTHORITY.OBSERVER ||
-                (user.getAuthority() == AUTHORITY.FR_RESPONSIBLE && project.getFRResp() != user)) {
+                (user.getAuthority() == AUTHORITY.FR_RESPONSIBLE && project.getFRResp().getId() != user.getId())) {
             return new ResponseEntity("You don't have premission to this resource", HttpStatus.UNAUTHORIZED);
         }
 
@@ -199,13 +214,45 @@ public class ProjectController {
 
     @PutMapping("/{projectId}/frteammembers/{memberId}")
     public ResponseEntity addFrTeamMember(@RequestHeader String googleTokenEncoded, @PathVariable Long projectId, @PathVariable Long memberId){
+        List<AUTHORITY> a = List.of(AUTHORITY.MODERATOR, AUTHORITY.ADMIN, AUTHORITY.FR_RESPONSIBLE);
+        String email = JwtVerifier.verifyAndReturnEmail(googleTokenEncoded);
+        if (email == null)
+            return new ResponseEntity("Token is missing or invalid", HttpStatus.UNAUTHORIZED);
+        if (userService.findByEmail(email) == null)
+            return new ResponseEntity("You don't have access to CDB", HttpStatus.UNAUTHORIZED);
+        if (!projectService.existsById(projectId))
+            return new ResponseEntity("Project not found", HttpStatus.NOT_FOUND);
+        if (!userService.existsById(memberId))
+            return new ResponseEntity("User not found", HttpStatus.NOT_FOUND);
+
+        AppUser endpointCaller = userService.findByEmail(email);
+        Project project = projectService.findById(projectId);
+        if (!a.contains(userService.findByEmail(email).getAuthority()) || project.getFRResp().getId() != endpointCaller.getId())
+            return new ResponseEntity("You don't have premission to this resource", HttpStatus.UNAUTHORIZED);
+
         projectService.addFrTeamMember(projectId, memberId);
         return new ResponseEntity("FR member added successfully", HttpStatus.OK);
     }
 
     @DeleteMapping("/{projectId}/frteammembers/{memberId}")
     public ResponseEntity deleteFrTeamMember(@RequestHeader String googleTokenEncoded, @PathVariable Long projectId, @PathVariable Long memberId){
+        List<AUTHORITY> a = List.of(AUTHORITY.MODERATOR, AUTHORITY.ADMIN, AUTHORITY.FR_RESPONSIBLE);
+        String email = JwtVerifier.verifyAndReturnEmail(googleTokenEncoded);
+        if (email == null)
+            return new ResponseEntity("Token is missing or invalid", HttpStatus.UNAUTHORIZED);
+        if (userService.findByEmail(email) == null)
+            return new ResponseEntity("You don't have access to CDB", HttpStatus.UNAUTHORIZED);
+        if (!projectService.existsById(projectId))
+            return new ResponseEntity("Project not found", HttpStatus.NOT_FOUND);
+        if (!userService.existsById(memberId))
+            return new ResponseEntity("User not found", HttpStatus.NOT_FOUND);
+
+        AppUser endpointCaller = userService.findByEmail(email);
+        Project project = projectService.findById(projectId);
+        if (!a.contains(userService.findByEmail(email).getAuthority()) || project.getFRResp().getId() != endpointCaller.getId())
+            return new ResponseEntity("You don't have premission to this resource", HttpStatus.UNAUTHORIZED);
+
         projectService.deleteFrTeamMember(projectId, memberId);
-        return new ResponseEntity("FR member added successfully", HttpStatus.OK);
+        return new ResponseEntity("FR member deleted successfully", HttpStatus.OK);
     }
 }
